requirements.txt
Django~=4.1.7
requests~=2.28.2
python-dotenv~=1.0.0